+++ src/test/java/com/birca/bircabackend/command/like/application/BirthdayCafeLikeServiceTest.java
@@ -1,7 +1,7 @@
-package com.birca.bircabackend.command.birca.application;
+package com.birca.bircabackend.command.like.application;
 
 import com.birca.bircabackend.command.auth.authorization.LoginMember;
-import com.birca.bircabackend.command.birca.domain.BirthdayCafeLike;
+import com.birca.bircabackend.command.like.domain.BirthdayCafeLike;
 import com.birca.bircabackend.command.birca.exception.BirthdayCafeErrorCode;
 import com.birca.bircabackend.common.exception.BusinessException;
 import com.birca.bircabackend.support.enviroment.ServiceTest;
